+++ packages/jest-rut/src/index.ts
@@ -3,8 +3,8 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
-import { matchers, unmockFetches, MatchResult } from 'rut';
-import { integrate } from 'rut/lib/configure';
+import { unmockFetches, MatchResult } from 'rut';
+import { integrate, matchers } from 'rut/lib/integrations';
 import * as integration from './integration';
 import serializer from './serializer';
