+++ packages/rut-dom/README.md
@@ -1,0 +1,15 @@
+# Jest + Rut
+
+[![Build Status](https://travis-ci.org/milesj/rut.svg?branch=master)](https://travis-ci.org/milesj/rut)
+[![npm version](https://badge.fury.io/js/jest-rut.svg)](https://www.npmjs.com/package/jest-rut)
+[![npm deps](https://david-dm.org/milesj/rut.svg?path=packages/jest-rut)](https://www.npmjs.com/package/jest-rut)
+
+Provides built-in matchers and serializers for testing Rut render results.
+
+## Requirements
+
+- Jest 24.0+
+
+## Documentation
+
+[https://milesj.gitbook.io/rut/setup#jest](https://milesj.gitbook.io/rut/setup#jest)
