+++ packages/rut/src/Element.ts
@@ -3,29 +3,24 @@
 import React from 'react';
 import { ReactTestInstance } from 'react-test-renderer';
 import {
-  InferEventFromHandler,
-  InferComponentProps,
-  HostComponentType,
   Predicate,
   DispatchOptions,
   DebugOptions,
   UnknownProps,
   AtIndexType,
-  EventType,
-  EventOptions,
-  InferHostElement,
-  EventMap,
+  QueryOptions,
 } from './types';
 import { getTypeName } from './helpers';
 import { doAct, doAsyncAct } from './internals/act';
 import { debug } from './internals/debug';
 import { getPropForDispatching } from './internals/element';
-import { whereTypeAndProps } from './predicates';
+import { whereTypeAndProps } from './helpers/predicates';
 import { factorySyntheticEvent } from './mocks/event';
 
-export default class Element<
+export default abstract class Element<
   Type extends React.ElementType = React.ElementType,
-  Host = InferHostElement<Type>
+  Props = UnknownProps,
+  Host = unknown
 > {
   protected element: ReactTestInstance;
 
@@ -50,15 +45,9 @@ export default class Element<
    * Dispatch an event listener for the defined prop name. Requires a synthetic event
    * based on the original event type.
    *
-   * Note: This may only be executed on host components (DOM elements).
+   * Note: This may only be executed on host components.
    */
-  dispatch<K extends EventType>(
-    name: K,
-    eventOrConfig?:
-      | InferEventFromHandler<EventMap<Host>[K]>
-      | EventOptions<Host, InferEventFromHandler<EventMap<Host>[K]>>,
-    options: DispatchOptions = {},
-  ): this {
+  dispatch(name: string, eventOrConfig?: unknown, options: DispatchOptions = {}): this {
     const prop = getPropForDispatching(this, name);
     const event = factorySyntheticEvent(name, eventOrConfig, this.element.type);
 
@@ -76,16 +65,13 @@ export default class Element<
   /**
    * Like `dispatch` but also awaits the event so that async calls have time to finish.
    */
-
-  async dispatchAndWait<K extends EventType>(
-    name: K,
-    eventOrConfig?:
-      | InferEventFromHandler<EventMap<Host>[K]>
-      | EventOptions<Host, InferEventFromHandler<EventMap<Host>[K]>>,
+  async dispatchAndWait(
+    name: string,
+    eventOrConfig?: unknown,
     options: DispatchOptions = {},
   ): Promise<void> {
     const prop = getPropForDispatching(this, name);
-    const event = factorySyntheticEvent(name, eventOrConfig, this.element.type);
+    const event = this.createSyntheticEvent(name, eventOrConfig, this.element.type);
 
     // istanbul ignore next
     if (options.propagate) {
@@ -100,14 +86,6 @@ export default class Element<
    * Search through the current tree for all elements that match the defined React
    * component or HTML type. If any are found, a list of `Element`s is returned.
    */
-  find<T extends HostComponentType, P extends InferComponentProps<T>>(
-    type: T,
-    props?: Partial<P>,
-  ): Element<T>[];
-  find<T extends React.ComponentType<any>, P extends InferComponentProps<T>>(
-    type: T,
-    props?: Partial<P>,
-  ): Element<T>[];
   find(type: React.ElementType<unknown>, props?: UnknownProps): Element<React.ElementType>[] {
     return this.query(whereTypeAndProps(type, props));
   }
@@ -117,16 +95,6 @@ export default class Element<
    * component or HTML type. If any are found, return the `Element` at the defined
    * index. Accepts shorthand `first` and `last` indices.
    */
-  findAt<T extends HostComponentType, P extends InferComponentProps<T>>(
-    type: T,
-    at: AtIndexType,
-    props?: Partial<P>,
-  ): Element<T>;
-  findAt<T extends React.ComponentType<any>, P extends InferComponentProps<T>>(
-    type: T,
-    at: AtIndexType,
-    props?: Partial<P>,
-  ): Element<T>;
   findAt(
     type: React.ElementType<unknown>,
     at: AtIndexType,
@@ -161,14 +129,6 @@ export default class Element<
    * component or HTML type. If exactly 1 is found, a `Element`s is returned,
    * otherwise an error is thrown.
    */
-  findOne<T extends HostComponentType, P extends InferComponentProps<T>>(
-    type: T,
-    props?: Partial<P>,
-  ): Element<T>;
-  findOne<T extends React.ComponentType<any>, P extends InferComponentProps<T>>(
-    type: T,
-    props?: Partial<P>,
-  ): Element<T>;
   findOne(type: React.ElementType<unknown>, props?: UnknownProps): Element<React.ElementType> {
     const results = this.find(type, props);
 
@@ -196,13 +156,10 @@ export default class Element<
    * ReactTestRenderer node and internal React fiber node. If any are found,
    * a list of `Element`s is returned.
    */
-  query<T extends React.ElementType>(
-    predicate: Predicate,
-    options?: { deep?: boolean },
-  ): Element<T>[] {
+  query<T extends React.ElementType>(predicate: Predicate, options?: QueryOptions): Element<T>[] {
     return this.element
       .findAll(node => predicate(node, node._fiber), { deep: true, ...options })
-      .map(node => new Element(node));
+      .map(node => new (this.constructor as any)(node));
   }
 
   /**
@@ -237,4 +194,13 @@ export default class Element<
   toString(): string {
     return this.name(true);
   }
+
+  /**
+   * Create a synthetic event specific to this renderer.
+   */
+  abstract createSyntheticEvent(
+    name: string,
+    eventOrConfig: unknown,
+    type: React.ElementType,
+  ): React.SyntheticEvent;
 }
