+++ packages/rut/src/integrations.ts
@@ -1,0 +1,4 @@
+import { integrate } from './internals/config';
+import matchers from './matchers';
+
+export { integrate, matchers };
