+++ packages/rut/src/types.ts
@@ -1,6 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any, @typescript-eslint/no-namespace */
 
 import React from 'react';
+import { ReactTestInstance } from 'react-test-renderer';
 import RutElement from './Element';
 import RutResult from './Result';
 
@@ -13,6 +14,11 @@ export interface RendererOptions {
   strict?: boolean;
   /** Wraps the root element in the provided React element. */
   wrapper?: React.ReactElement;
+}
+
+export interface AdapterRendererOptions extends RendererOptions {
+  /** Wrap a react test instance with a Rut element. */
+  createElement: (instance: ReactTestInstance) => RutElement;
 }
 
 export interface DebugOptions {
@@ -43,9 +49,9 @@ export interface DispatchOptions {
   propagate?: boolean;
 }
 
-export interface GlobalOptions extends RendererOptions {
-  /** React renderer that is being tested. */
-  renderer?: 'dom';
+export interface QueryOptions {
+  /** Search through the entire React trender tree. */
+  deep?: boolean;
 }
 
 export interface IntegrationOptions {
@@ -111,69 +117,13 @@ export type NodeType =
   | 'host-component'
   | 'memo';
 
-// COMPONENTS
-
-export type HostComponentType = keyof JSX.IntrinsicElements;
-
-export type InferHostElement<T> = T extends keyof HTMLElementTagNameMap
-  ? HTMLElementTagNameMap[T]
-  : T extends keyof SVGElementTagNameMap
-  ? SVGElementTagNameMap[T]
-  : unknown;
-
-export type InferComponentProps<T> = T extends HostComponentType
-  ? JSX.IntrinsicElements[T]
-  : T extends React.ComponentType<infer P>
-  ? P
-  : {};
-
-// EVENTS
-
-export type InferEventFromHandler<T> = T extends (event: infer E) => void ? E : never;
-
-export type InferHostElementFromEvent<T> = T extends React.SyntheticEvent<infer E> ? E : Element;
-
-export type EventMap<T> = Required<
-  Omit<React.DOMAttributes<T>, 'children' | 'dangerouslySetInnerHTML'>
->;
-
-export type EventType = keyof EventMap<unknown>;
-
-export type EventOptions<T, E> = {
-  currentTarget?: Partial<T>;
-  target?: Partial<T>;
-} & ExpandedEventOptions<E>;
-
-export type ExpandedEventOptions<T> = T extends React.AnimationEvent | AnimationEvent
-  ? { animationName?: string }
-  : T extends
-      | React.MouseEvent
-      | React.KeyboardEvent
-      | React.TouchEvent
-      | MouseEvent
-      | KeyboardEvent
-      | TouchEvent
-  ? {
-      altKey?: boolean;
-      ctrlKey?: boolean;
-      key?: string;
-      keyCode?: number;
-      metaKey?: boolean;
-      shiftKey?: boolean;
-    }
-  : T extends React.TransitionEvent | TransitionEvent
-  ? { propertyName?: string }
-  : {};
-
 // AUGMENTATION
 
 export type PropsOf<T> = T extends RutResult<infer P>
   ? P
-  : T extends RutElement<infer E>
-  ? InferComponentProps<E>
+  : T extends RutElement<any, infer P>
+  ? P
   : {};
-
-export type StructureOf<T> = { [K in keyof T]: T[K] };
 
 declare module 'react-test-renderer' {
   interface ReactTestRenderer {
